--- a/base/loading.jl
+++ b/base/loading.jl
@@ -636,10 +636,14 @@ function find_all_in_cache_path(pkg::PkgId)
             isfile_casesensitive(filepath) && push!(paths, filepath)
         end
     end
-    # allocating the sort vector is less expensive than using sort!(.. by=mtime), which would
-    # call the relatively slow mtime multiple times per path
-    p = sortperm(mtime.(paths), rev = true)
-    return paths[p]
+    if length(paths) > 1
+        # allocating the sort vector is less expensive than using sort!(.. by=mtime), which would
+        # call the relatively slow mtime multiple times per path
+        p = sortperm(mtime.(paths), rev = true)
+        return paths[p]
+    else
+        return paths
+    end
 end
 
 # these return either the array of modules loaded from the path / content given