@KristofferC 2020-11-03
Nice, I think it might be better to try the path from `compilecache_path` if it exists first, then go in modified order.