@KristofferC 2020-11-03
Actually, I think the new preference system makes that impossibe... cf the comment:

>  # Read preferences hash back from .ji file (we can't precompute because
>  # we don't actually know what the list of compile-time preferences are without compiling)

That's kind of unfortunate. Is there no way to compute this without actually doing the compilation? cc @staticfloat