@ianshmean 2020-11-03
One more test:

### Master
```
julia> @time using DifferentialEquations
  6.861824 seconds (16.32 M allocations: 1.242 GiB, 6.40% gc time, 33.17% compilation time)
```
with 1,302 `open` calls and 20,306 `stat64` calls

### This PR
```
julia> @time using DifferentialEquations
  6.502538 seconds (13.12 M allocations: 1.036 GiB, 4.33% gc time, 26.88% compilation time)
```
with 620 `open` calls and 24,959 `stat64` calls