@staticfloat 2020-11-03
> That's kind of unfortunate. Is there no way to compute this without actually doing the compilation?

That's right; unfortunately, there's no way to know what set of preferences were accessed during compilation without actually running the code.  We record the set within the `.ji` file, but clearly that's not helpful here.

While at first I was surprised that this is taking so long, now that I think about it, it makes perfect sense.  We've got batches of serialized, asynchronous calls (e.g. "open this file, read some contents, if it's not the one we want, load another") all serialized after each other ("load this package, then once that's done, load this other package").  I wonder what would happen if we could check the files in parallel, much less load packages in parallel.