@KristofferC 2020-11-03
>  I wonder what would happen if we could check the files in parallel, much less load packages in parallel.

Yeah, that's an interesting idea.