@ianshmean 2020-11-04
I  realised that the `sort!(..  by=..` was calling `mtime` more than once on  each file, and given it's expensive it's faster to allocate a sort vector