@ianshmean 2020-11-06
Perhaps, would it be appropriate to assume that most packages won't have custom preferences set up, so we could try the null preferences slug (which is stable with project & julia build) before falling back to search for other caches to try? I implemented that on a branch here, based on this PR https://github.com/ianshmean/julia/compare/ib/cache_file_order...ianshmean:ib/cache_file_order_check_null_first

I did wonder whether the active project needs to be encoded into the slug, given the preferences and julia build is already? With it included, the slug will need to be re-calculated when the active project changes, so will need a call to be added to `Pkg.activate`