@ianshmean 2020-11-23
I believe so. There is the open question about my comment above, but if that is worth exploring it could be a future PR