@KristofferC 2020-11-24
Some weird download error on macos, unrelated to this PR but not sure if the error is something real or because the branch is perhaps not rebased recently, cc @StefanKarpinski 

``` 
Error During Test at /Users/julia/buildbot/worker-tabularasa/tester_macos64/build/share/julia/stdlib/v1.6/Downloads/test/runtests.jl:184
  Got exception outside of a @test
  TaskFailedException
  
      nested task error: Couldn't connect to server while requesting https://httpbingo.julialang.org/delay/2?id=6
      Stacktrace:
        [1] (::Downloads.var"#9#17"{IOBuffer, Base.DevNull, Nothing, Vector{Pair{String, String}}, Nothing, Bool, Bool, String, Int64, Bool, Bool})(easy::Downloads.Curl.Easy)
          @ Downloads ~/buildbot/worker/package_macos64/build/usr/share/julia/stdlib/v1.6/Downloads/src/Downloads.jl:313
...
```